+++ activiti-cloud-runtime-bundle-service/activiti-cloud-services-runtime-bundle/activiti-cloud-services-connectors/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,4 +1,3 @@
 org.activiti.services.connectors.conf.CloudConnectorsAutoConfiguration
 org.activiti.services.connectors.conf.CloudConnectorsMessagingAutoConfiguration
-org.activiti.services.connectors.conf.RuntimeBundleShedLockConfiguration
 org.activiti.services.connectors.recovery.OrphanedIntegrationRecoveryConfiguration
